requirements.txt
Flask
python-dotenv
googletrans==4.0.0-rc1
SpeechRecognition
gTTS
moviepy
pydub
ffmpeg-python
